feat() adiciona endpoint para updlod de imagem do produto

+++ src/main/java/br/com/fiap/techchallenge/lanchonete/adapters/repository/jpa/ProdutoJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProdutoJpaRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoJpaRepository extends JpaRepository<Produto, Long> {
 }
